created update booking route

+++ src/modules/vehicle/vehicle.service.ts
@@ -7,11 +7,9 @@ const addVehicle = async (vehicle_name: string, type: string,registration_number
     );
     return result
 }
-                         /* working here  */
-                         
-const getVehicles = async (role: string) => {
-const result = role == "admin" ? await pool.query(`SELECT * FROM vehicle`) : await pool.query(`
-    SELECT id, vehicle_id, rent_start_date,rent_end_date,total_price,status WHERE id=$1 FROM booking`)
+                                                          
+const getVehicles = async () => {
+const result = await pool.query(`SELECT * FROM vehicle`)
 return result
 }
 
@@ -33,8 +31,20 @@ const updateVehicle =async (vehicle_name :string, type:string, registration_numb
 } 
 
 const deleteVehicle = async (id :string) => {
-     const result = await pool.query(`DELETE FROM vehicle WHERE id=$1`, [id]);
-    return result
+   const vehicle = await pool.query(`SELECT availability_status FROM vehicle WHERE id=$1`,[id])
+
+    if (vehicle.rowCount === 0){
+    throw new Error ("vahicle not found")
+   }
+    
+   const status = vehicle.rows[0].availability_status
+   
+   if (status === "booked"){
+     throw new Error ("Vehicle is booked. Cannot delete")
+   }
+   const deleteResult = await pool.query(`DELETE FROM vehicle WHERE id=$1`,[id])
+   
+   return deleteResult
 }
 
 export const vehicleService = {
